VIC-13106 Fix for UK alarms bug (#2)
Setting a timer with the device timezone set to Europe/London results
in a 1 hour delay. This is because the SDK's approach to converting
from a std::tm to std::time_t can be tricked by daylight savings time.
Common internet suggestions are to use timegm (GNU'd), set the TZ env
variable prior to calling mktime (not thread safe), or doing the math
manually.
This uses some code from Quickstep to do the conversion manually.

The bug is tracked in avs-device-sdk issue 1183, but no solution is
provided yet.